requirements.txt
openai
python-dotenv
pydantic
streamlit
reportlab